Cleanup for Recommended: IrqlSetTooLow
